Disable buffering for demo mode.
If the database is not connected, a "Not Connected to a Database" message is shown at the top of the Song Database list in the Producer View, as well as extra instructions if demoMode is active.

Added demoMode option which toggles demo json data when the database is not connected.

The browser is automatically opened to the localhost port when `npm run start` (node server.js) is executed.